git-svn-id: https://matrix.tplinkdns.com:8443/svn/MyCode/VNet/VNet.ProceduralGeneration@1324 0f8fcb7d-4e12-9a4c-a6b3-2570e6efd5fd

+++ VNet.ProceduralGeneration/Cosmological/Contexts/ContextBase.cs
@@ -7,9 +7,9 @@ namespace VNet.ProceduralGeneration.Cosmological.Contexts
     {
         public float Age { get; set; }                      // years
         public double Mass { get; set; }                    // kg
-        public float Size { get; set; }                     // AU
+        public double Size { get; set; }                     // AU
         public float AbsoluteMagnitude { get; set; }        // 
-        public double Luminosity { get; set; }              //  L⊙
+        public float Luminosity { get; set; }              //  L⊙
         public float Temperature { get; set; }              // Kelvin
         public Vector3 Position { get; set; }               // AU
